Initial Impact of Telepharmacy on  Specialty Medication Adherence
Patients with specialty conditions are faced 
with a unique set of challenges in regard 
to being adherent to their medications.1
One-on-one pharmacist counseling from 
the comfort of the patient’s home can be 
an innovative and impactful intervention 
that can lead to improved adherence for 
specialty patients. This study analyzes the 
impact on adherence after an initial video 
conferencing session between pharmacist 
and patient at the start of treatment.